REFACTOR: Bluetooth functionality refactored

+++ persefone/app/src/main/java/se/ju/lejo/persefone/Data/DataHandler.kt
@@ -3,7 +3,7 @@ package se.ju.lejo.persefone.Data
 object DataHandler {
 
     private var _isClockedIn: Boolean = false
-    var _E: Long = 30
+    var _E: Float = 30.0f
 
     fun setIsClockedIn(isClockedIn: Boolean) {
         this._isClockedIn = isClockedIn
@@ -13,11 +13,11 @@ object DataHandler {
         return this._isClockedIn
     }
 
-    fun setE(e: Long) {
+    fun setE(e: Float) {
         this._E = e
     }
 
-    fun getE(): Long {
+    fun getE(): Float {
         return this._E
     }
 }
